requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
pyqt5>=5.15.0
pyserial>=3.5
pyqtgraph>=0.12.0
pyopengl>=3.1.5
plotly>=5.3.0 